main: Suppress MinGW warnings
Cast `off_t` to `long long` and use "%lld" for the format string.

The size of `off_t` is 32-bit by default on Win32 (VC and MinGW),
however it becomes 64-bit when `_FILE_OFFSET_BITS` is defined to 64 on
MinGW-w64 (both 32-bit and 64-bit). It's not so easy to set the format
string properly. Using `long long` is the easiest way.

+++ main/lregex.c
@@ -25,8 +25,6 @@
 #endif
 #include <regex.h>
 
-#include <inttypes.h>
-
 #include "debug.h"
 #include "colprint_p.h"
 #include "entry_p.h"
@@ -2755,10 +2753,10 @@ static void   guestRequestSubmit (struct guestRequest *r)
 {
 	const char *langName = getLanguageName (r->lang);
 	verbose ("guestRequestSubmit: %s; "
-			 "range: %"PRIdPTR" - %"PRIdPTR"\n",
+			 "range: %lld - %lld\n",
 			 langName,
-			 r->boundary[BOUNDARY_START].offset,
-			 r->boundary[BOUNDARY_END].offset);
+			 (long long)r->boundary[BOUNDARY_START].offset,
+			 (long long)r->boundary[BOUNDARY_END].offset);
 	makePromiseForAreaSpecifiedWithOffsets (langName,
 											r->boundary[BOUNDARY_START].offset,
 											r->boundary[BOUNDARY_END].offset);
